break up CommandGift
